shop v.7.2.4 New features: * Option to select a courier in any order status by means of new action “Edit shipping details”. * Editing of order shipping date and time in backend. * Order sorting in backend by shipping date and time, and by status. * Option not to send follow-ups if selected order status has changed. Fixed: * Reports: displaying of order list for manually selected time periods. * Category filters: search of products with price specified in currencies different from that selected by a customer in storefront. * Category filters: automatic update of specified price limit values on currency change in storefront. * Storefront settings: added “og:url” meta tag field. * Order editing: displaying of an error message if order contains a deleted SKU.

+++ lib/actions/settings/followups/shopSettingsFollowups.action.php
@@ -131,8 +131,8 @@ class shopSettingsFollowupsAction extends waViewAction
         $this->view->assign('followup', $followup);
         $this->view->assign('followups', $followups);
         $this->view->assign('test_orders', $test_orders);
-        $this->view->assign('last_cron', wa()->getSetting('last_followup_cli'));
-        $this->view->assign('cron_ok', wa()->getSetting('last_followup_cli') + 3600*36 > time());
+        $this->view->assign('last_cron', (int)wa()->getSetting('last_followup_cli'));
+        $this->view->assign('cron_ok', ((int)wa()->getSetting('last_followup_cli') + 3600*36) > time());
         $this->view->assign('cron_command', 'php '.wa()->getConfig()->getRootPath().'/cli.php shop followup');
         $this->view->assign('default_email_from', $this->getConfig()->getGeneralSettings('email'));
         $this->view->assign('routes', wa()->getRouting()->getByApp('shop'));
